refactor : repository 수정
